Add sub-category pages and service and related methods

+++ HomeService.Infra.DataAccsess.Repos.EF/Repositories/SuggestionRepository.cs
@@ -64,14 +64,14 @@ public class SuggestionRepository : ISuggestionRepository
     public async Task AcceptSuggestion(int id,CancellationToken cancellationToken)
     {
       var targetSuggestion=  await _context.Suggestions.FirstOrDefaultAsync(s => s.Id == id,cancellationToken);
-        targetSuggestion.Status =(int)StatusEnum.Confirmed;
+        targetSuggestion.Status =StatusEnum.Confirmed;
 
         _context.SaveChangesAsync(cancellationToken);
     }
 
     public async Task<int> ConfrimedStatusCount(int orderId,CancellationToken cancellationToken)
     {
-      return await _context.Suggestions.Where(s =>s.OrderId==orderId && s.Status ==(int) StatusEnum.Confirmed).CountAsync(cancellationToken);
+      return await _context.Suggestions.Where(s =>s.OrderId==orderId && s.Status == StatusEnum.Confirmed).CountAsync(cancellationToken);
     }
     
     private async Task<Suggestion> FindSuggestion(int id, CancellationToken cancellationToken)
